feat: 사진 선택 10개로 제한 및 토스트 메세지 추가
